+++ lib/onvif.dart
@@ -4,8 +4,10 @@ library easy_onvif;
 export 'src/onvif_base.dart';
 
 export 'src/device_management.dart';
+export 'src/imaging.dart';
 export 'src/media.dart';
 export 'src/multicast_probe.dart';
 export 'src/ptz.dart';
 export 'src/recordings.dart';
 export 'src/replay.dart';
+export 'src/search.dart';