+++ lib/soap.dart
@@ -3,11 +3,13 @@ library soap;
 
 export 'src/soap/authorization.dart';
 export 'src/soap/device_management.dart';
+export 'src/soap/imaging.dart';
 export 'src/soap/media1.dart';
 export 'src/soap/media2.dart';
 export 'src/soap/nonce.dart';
 export 'src/soap/ptz.dart';
 export 'src/soap/recordings.dart';
 export 'src/soap/replay.dart';
+export 'src/soap/search.dart';
 export 'src/soap/transport.dart';
 export 'src/soap/xmlns.dart';