+++ lib/src/soap/xmlns.dart
@@ -3,12 +3,14 @@ class Xmlns {
   static const s = 'http://www.w3.org/2003/05/soap-envelope';
   static const tds = 'http://www.onvif.org/ver10/device/wsdl';
   static const ter = 'http://www.onvif.org/ver10/error';
+  static const timg = 'http://www.onvif.org/ver20/imaging/wsdl';
   static const tptz = 'http://www.onvif.org/ver20/ptz/wsdl';
   static const tns1 = 'http://www.onvif.org/ver10/topics/';
   static const trt = 'http://www.onvif.org/ver10/media/wsdl';
   static const tr2 = 'http://www.onvif.org/ver20/media/wsdl';
   static const trc = 'http://www.onvif.org/ver10/recording/wsdl';
   static const trp = 'http://www.onvif.org/ver10/replay/wsdl';
+  static const tse = 'http://www.onvif.org/ver10/search/wsdl';
   static const tt = 'http://www.onvif.org/ver10/schema';
   static const xsd = 'http://www.w3.org/2001/XMLSchema';
 }